+++ libp2p/src/lib.rs
@@ -123,17 +123,17 @@ pub use libp2p_uds as uds;
 #[cfg(not(target_arch = "wasm32"))]
 #[doc(inline)]
 pub use libp2p_upnp as upnp;
-#[cfg(feature = "webrtc-websys")]
+#[cfg(all(feature = "webrtc-websys", target_arch = "wasm32"))]
 #[doc(inline)]
 pub use libp2p_webrtc_websys as webrtc_websys;
 #[cfg(feature = "websocket")]
 #[cfg(not(target_arch = "wasm32"))]
 #[doc(inline)]
 pub use libp2p_websocket as websocket;
-#[cfg(feature = "websocket-websys")]
+#[cfg(all(feature = "websocket-websys", target_arch = "wasm32"))]
 #[doc(inline)]
 pub use libp2p_websocket_websys as websocket_websys;
-#[cfg(feature = "webtransport-websys")]
+#[cfg(all(feature = "webtransport-websys", target_arch = "wasm32"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "webtransport-websys")))]
 #[doc(inline)]
 pub use libp2p_webtransport_websys as webtransport_websys;