+++ libp2p/CHANGELOG.md
@@ -2,6 +2,9 @@
 - Introduce `libp2p-webrtc-websys` behind `webrtc-websys` feature flag.
   See [PR 5819](https://github.com/libp2p/rust-libp2p/pull/5819).
 
+- Make the `*-websys` variants (`libp2p-webrtc-websys`, `libp2p-websocket-websys`, `libp2p-webtransport-websys`) only available on wasm32 target architecture.
+  See [PR 5891](https://github.com/libp2p/rust-libp2p/pull/5891).
+  
 - Adds TopicsConfig structure and topic_configuration key to Config and ProtocolConfig to enable
   setting message and mesh-*-n sizes on the Gossipsub configuration. Adds appropriate functions to
   allow modifying default values from the config interface.